be/requirements-ai.txt
fastapi==0.116.1
openai==2.6.1
python-dotenv==1.1.1
uvicorn==0.35.0
